+++ out/corvus.js
@@ -18,32 +18,51 @@ program
     .command('check')
     .alias('c')
     .description('Check dependencies.')
-    .action(function () { return console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('check'), 'is not yet ready for usage.'); });
+    .action(function () {
+    tasker.check();
+    console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('check'), 'is not yet ready for usage.');
+});
 program
     .command('install')
     .alias('i')
     .description('Install dependencies.')
-    .action(function () { return console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('install'), 'is not yet ready for usage.'); });
+    .action(function () {
+    tasker.install();
+    console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('install'), 'is not yet ready for usage.');
+});
 program
-    .command('generate <type> [attribute]')
+    .command('generate <type> [frameworks...]')
     .alias('g')
     .description('Scaffolding')
-    .action(function () { return console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('generate'), 'is not yet ready for usage.'); });
+    .action(function (type, frameworks) {
+    tasker.generate(type, frameworks);
+    if (frameworks) {
+        frameworks.forEach(function (framework) {
+            console.log(framework);
+        });
+    }
+    console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('generate'), 'is not yet ready for usage.');
+});
 program
     .command('revert')
     .alias('r')
     .description('Revert to initial creation')
-    .action(function () { return console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('revert'), 'is not yet ready for usage.'); });
+    .action(function () {
+    tasker.revert();
+    console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('revert'), 'is not yet ready for usage.');
+});
 program
     .command('p <command>')
-    .alias('ps')
     .description('Git push')
-    .action(function (command) { return console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold(command), 'is not yet ready for usage.'); });
+    .action(function (command) {
+    console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold(command), 'is not yet ready for usage.');
+});
 program
     .command('add <name> <type> <val>')
     .alias('a')
     .description('Add a new field')
     .action(function (name, type, val) {
+    tasker.add(name, type, val);
     var t;
     switch (type) {
         case "str" || "string":
@@ -62,3 +81,6 @@ program
     console.log(chalk.cyan.bold('CORVUS'), chalk.red.bold('ERR!'), chalk.magenta.bold('add'), 'is not yet ready for usage.');
 });
 program.parse(process.argv);
+if (!process.argv.slice(2).length) {
+    program.outputHelp();
+}